[/]: Projects: split the component to smaller components
