+++ components/pages/Projects/ProjectContainer/ProjectContainer.component.tsx
@@ -2,8 +2,8 @@
 import style from "./ProjectContainer.component.module.scss";
 
 import Image from "next/image";
-import AnimationWrapper from "../AnimationWrapper/AnimationWrapper.component";
-import { ProjectType } from "@/types/PostType.type";
+import AnimationWrapper from "../../../UI/AnimationWrapper/AnimationWrapper.component";
+import { type ProjectType } from "@/types/PostType.type";
 
 const ProjectComponent = ({
   image,
